Bug fix: Constant bitrate of transport stream.

+++ include/project/si/PMTView.h
@@ -40,7 +40,7 @@ class PMTView : public ProjectInfo {
 		unsigned short programNumber;
 		int tsinput;
 		unsigned short pcrPid;
-		unsigned short pcrFrequency;
+		unsigned int pcrFrequency;
 		string serviceName;
 		short int serviceType;
 		ProjectInfo* eitProj;
@@ -65,8 +65,8 @@ class PMTView : public ProjectInfo {
 		int getTsinput();
 		void setPcrPid(unsigned short pid);
 		unsigned short getPcrPid();
-		void setPcrFrequency(unsigned short freq);
-		unsigned short getPcrFrequency();
+		void setPcrFrequency(unsigned int freq);
+		unsigned int getPcrFrequency();
 		void setServiceName(string name);
 		string getServiceName();
 		void setServiceType(short int st);
